Make sure connections are aborted correctly on Windows
Fixes the first failure mentioned in #19611, 'must shut down both streams
when connection is aborted remotely'.

This might not be the proper fix yet but we could work from here.

The problem was that when aborting, SO_LINGER is enabled, which causes
Windows not to close the connection on preClose but on kill.

However, when there's selectors registered to the channel (even if they
haven't expressed interest in any actual events), the channel is not
immediately killed, but marked as cancelled. Cancelled connections are
killed around select(), which is the reason waking up the select()
triggers closing the connection here.